added static files and  function for work with files, updated posts and users functionals

+++ src/posts/postService.js
@@ -1,9 +1,12 @@
 import postSchema from "./postSchema.js"
+import fileService from "../fileService.js"
 
 class PostService {
-  async createPost(post) {
+  async createPost(post, picture) {
+    const fileName = fileService.saveFile(picture)
+
     try {
-      const createdPost = await postSchema.create(post)
+      const createdPost = await postSchema.create({...post, picture: fileName})
       await createdPost.save()
       return createdPost
     } catch (e) {
@@ -22,7 +25,6 @@ class PostService {
 
   async getOnePost(_id) {
     try {
-      console.log(_id)
       const post = await postSchema.findOne({_id})
       return post
     } catch (e) {
@@ -30,9 +32,12 @@ class PostService {
     }
   }
 
-  async updatePost(_id, data) {
+  async updatePost(_id, data, picture) {
+    if (!_id) throw new Error('ID not found')
+
     try {
       const updatedPost = await postSchema.findOneAndUpdate({_id}, data, {new: true})
+      if (picture) await fileService.updateFile(picture, updatedPost.picture)
       return updatedPost
     } catch (e) {
       throw new Error(e)
@@ -40,8 +45,11 @@ class PostService {
   }
 
   async deletePost(_id) {
+    if (!_id) throw new Error('ID not found')
+
     try {
       const deletedPost = await postSchema.findOneAndDelete({_id})
+      await fileService.deleteFile(deletedPost.picture)
       return deletedPost
     } catch (e) {
       throw new Error(e)
